make src/core.mjs tree-shakeable, use it as source of truth
