+++ packages/joint-core/src/core.mjs
@@ -1,48 +1,25 @@
-import * as connectors from './connectors/index.mjs';
-import * as highlighters from './highlighters/index.mjs';
-import * as connectionPoints from './connectionPoints/index.mjs';
-import * as connectionStrategies from './connectionStrategies/index.mjs';
-import * as routers from './routers/index.mjs';
-import * as anchors from './anchors/index.mjs';
-import * as linkAnchors from './linkAnchors/index.mjs';
-import * as dia from './dia/index.mjs';
-import * as linkTools from './linkTools/index.mjs';
-import * as elementTools from './elementTools/index.mjs';
-import * as util from './util/index.mjs';
-import * as mvc from './mvc/index.mjs';
-import * as g from './g/index.mjs';
-import { config } from './config/index.mjs';
 import V from './V/index.mjs';
-import * as Port from './layout/ports/port.mjs';
-import * as PortLabel from './layout/ports/portLabel.mjs';
 
-export * from '../dist/version.mjs';
+export * as connectors from './connectors/index.mjs';
+export * as highlighters from './highlighters/index.mjs';
+export * as connectionPoints from './connectionPoints/index.mjs';
+export * as connectionStrategies from './connectionStrategies/index.mjs';
+export * as routers from './routers/index.mjs';
+export * as anchors from './anchors/index.mjs';
+export * as linkAnchors from './linkAnchors/index.mjs';
+export * as dia from './dia/index.mjs';
+export * as linkTools from './linkTools/index.mjs';
+export * as elementTools from './elementTools/index.mjs';
+export * as util from './util/index.mjs';
+export * as mvc from './mvc/index.mjs';
+export * as g from './g/index.mjs';
+export * as layout from './layout/index.mjs';
+
 export const Vectorizer = V;
-export const layout = { PortLabel, Port };
+export { V };
+
+export { config } from './config/index.mjs';
 export { env } from './env/index.mjs';
-export {
-    config,
-    anchors,
-    linkAnchors,
-    connectionPoints,
-    connectionStrategies,
-    connectors,
-    dia,
-    highlighters,
-    mvc,
-    routers,
-    util,
-    linkTools,
-    elementTools,
-    V,
-    g
-};
-export const setTheme = function(theme, opt) {
+export { setTheme } from './setTheme.mjs';
 
-    opt = opt || {};
-
-    util.invoke(mvc.views, 'setTheme', theme, opt);
-
-    // Update the default theme on the view prototype.
-    mvc.View.prototype.defaultTheme = theme;
-};
+export { version } from '../dist/version.mjs';
